% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupData.R
\name{setupData}
\alias{setupData}
\title{Setup data prior to use with Animal Tracking Toolbox functions}
\usage{
setupData(Tag.Detections, Tag.Metadata, Station.Information, tz = "UTC",
  crs = NULL)
}
\arguments{
\item{Tag.Detections}{data frame with tag detection data including coordinates. Currently only handles IMOS ATF data structure.}

\item{Tag.Metadata}{data frame with metadata for all tags represented in Tag.Detections. Currently only handles IMOS ATF metadata structure.}

\item{Station.Information}{data frame with information on receiver station including coordinates. Currently only handles IMOS ATF station information structure.}

\item{tz}{time zone of date time information in Tag.Detections, Tag.Metadata and Station.Information. If none provided defaults to "UTC"}

\item{crs}{geographic coordinate system for all Tag.Detections, Tag.Metadata and Station.Information (latitude/longitude). If none provided defaults to WGS84.}
}
\value{
Produces an 'ATT' object that is a list of tibbles containing Tag.Detections, Tag.Metadata and Station.Information.
The 'ATT' object will have a geographic coordinate system associated with it for smoother functioning of subsequent functions.
}
\description{
Sets up and stores all relevant information required for spatio-temporal analyses of passive telemetry data.
This function produces an 'ATT' object which standardises field names (currently handles IMOS ATF export data structure).
}
\examples{
## Import example datasets
data(tagdata)
data(taginfo)
data(statinfo)

## Setup data
ATTdata<- setupData(Tag.Detections = tagdata, Tag.Metadata = taginfo, Station.Information = statinfo)
ATTdata


}
\seealso{
setup data can be used to estimate detection \code{\link{detectionSummary}}, dispersal \code{\link{dispersalSummary}}
and Short-term center of activity \code{\link{COA}}.
}
